move web components and projects to subdirectory
