+++ src/index.ts
@@ -12,10 +12,8 @@ export * from './jsii-project';
 export * from './json';
 export * from './license';
 export * from './makefile';
-export * from './next';
 export * from './node-project';
 export * from './project';
-export * from './react';
 export * from './sample-file';
 export * from './semver';
 export * from './start';
@@ -23,9 +21,9 @@ export * from './textfile';
 export * from './toml';
 export * from './typescript';
 export * from './version';
-export * from './web';
 export * from './yaml';
 
 // export submobules
+export * as web from './web';
 export * as github from './github';
 export * as vscode from './vscode';
