+++ src/web/next.ts
@@ -1,10 +1,10 @@
-import { Component } from './component';
-import { FileBase, FileBaseOptions, IResolver } from './file';
-import { NodeProject, NodeProjectOptions } from './node-project';
-import { SampleDir } from './sample-file';
-import { StartEntryCategory } from './start';
-import { TypeScriptAppProject, TypeScriptJsxMode, TypeScriptModuleResolution, TypeScriptProjectOptions } from './typescript';
-import { PostCss } from './web';
+import { Component } from '../component';
+import { FileBase, FileBaseOptions, IResolver } from '../file';
+import { NodeProject, NodeProjectOptions } from '../node-project';
+import { SampleDir } from '../sample-file';
+import { StartEntryCategory } from '../start';
+import { TypeScriptAppProject, TypeScriptJsxMode, TypeScriptModuleResolution, TypeScriptProjectOptions } from '../typescript';
+import { PostCss } from './postcss';
 
 export interface NextJsCommonProjectOptions {
   /**
