+++ src/web/postcss.ts
@@ -1,5 +1,6 @@
-import { JsonFile } from './json';
-import { NodeProject } from './node-project';
+import { JsonFile } from '../json';
+import { NodeProject } from '../node-project';
+import { TailwindConfig, TailwindConfigOptions } from './tailwind';
 
 export interface PostCssOptions {
   /**
@@ -43,46 +44,6 @@ export class PostCss {
     }
 
     this.file = new JsonFile(project, this.fileName, { obj: config });
-
-    project.npmignore?.exclude(`/${this.fileName}`);
-  }
-}
-
-export interface TailwindConfigOptions {
-  /**
-   * @default "tailwind.config.json"
-   */
-  readonly fileName?: string;
-}
-
-/**
- * Declares a Tailwind CSS configuration file.
- *
- * There are multiple ways to add Tailwind CSS in your node project - see:
- * https://tailwindcss.com/docs/installation
- *
- * @see PostCss
- */
-export class TailwindConfig {
-  public readonly fileName: string;
-  public readonly file: JsonFile;
-
-  constructor(project: NodeProject, options?: TailwindConfigOptions) {
-    this.fileName = options?.fileName ?? 'tailwind.config.json';
-
-    this.file = new JsonFile(project, this.fileName, {
-      obj: {
-        purge: [],
-        darkMode: false,
-        theme: {
-          extend: {},
-        },
-        variants: {
-          extend: {},
-        },
-        plugins: [],
-      },
-    });
 
     project.npmignore?.exclude(`/${this.fileName}`);
   }
