+++ src/web/tailwind.ts
@@ -1,0 +1,42 @@
+import { JsonFile } from '../json';
+import { NodeProject } from '../node-project';
+
+export interface TailwindConfigOptions {
+  /**
+   * @default "tailwind.config.json"
+   */
+  readonly fileName?: string;
+}
+
+/**
+ * Declares a Tailwind CSS configuration file.
+ *
+ * There are multiple ways to add Tailwind CSS in your node project - see:
+ * https://tailwindcss.com/docs/installation
+ *
+ * @see PostCss
+ */
+export class TailwindConfig {
+  public readonly fileName: string;
+  public readonly file: JsonFile;
+
+  constructor(project: NodeProject, options?: TailwindConfigOptions) {
+    this.fileName = options?.fileName ?? 'tailwind.config.json';
+
+    this.file = new JsonFile(project, this.fileName, {
+      obj: {
+        purge: [],
+        darkMode: false,
+        theme: {
+          extend: {},
+        },
+        variants: {
+          extend: {},
+        },
+        plugins: [],
+      },
+    });
+
+    project.npmignore?.exclude(`/${this.fileName}`);
+  }
+}
